+++ src/pages/ProductList/components/SortBar/SortBar.js
@@ -1,30 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './SortBar.scss';
 import SortFilter from './SortFilter';
 
 const SortBar = () => {
-  const [onColor, setOnColor] = useState(false);
-
-  const switchFilter = () => {
-    setOnColor(prev => !prev);
-  };
-  const changeBackgroundColor = onColor => {
-    return onColor ? '#eee4d4' : '#9e887c';
-  };
   return (
     <div className="sortBar">
       <p>총 5개 상품이 있습니다</p>
       <ul>
         {LI_LIST.map(props => {
-          return (
-            <li
-              key={props.id}
-              onClick={switchFilter}
-              style={{ backgroundColor: `${changeBackgroundColor(onColor)}` }}
-            >
-              {props.title}
-            </li>
-          );
+          return <SortFilter title={props.title} key={props.id} />;
         })}
       </ul>
     </div>
